Add Artifact Upload Step to GitHub Workflow
This commit introduces a new step in the GitHub workflow to archive and upload the lint analysis reports. The lint report, lint-results-debug.html, is now uploaded as an artifact named 'Lint-Reports' after each workflow run. This allows for easier access and review of the lint reports directly from the GitHub interface.