Fix the language is not matching with Figma  (#2460)
* Update English language name in AppLanguage enum

Changed the mainName for EN from 'English UK' to 'English' and set engName to '(UK)' for consistency with other language entries.

* Update language mapping for English in AppLanguage

Changed the string comparison from 'English UK' to 'English' in the AppLanguage enum to ensure correct mapping for the English language.

* Add support for 'English UK' in AppLanguage enum

Updated the fromName() function to recognize 'English UK' as a valid input, mapping it to the EN language option.